Improve upload file function compatibility
The current upload function calculates the available space in the remote system using `df` and by specifying the block size using the `--block-size` flag. That flag exists only in GNU version of `df` but it fails with BusyBox or Toybox version of `df`.
The `-k` flag, which sets the block size to 1024 bytes, is preferred as common among different implementations of `df`.